requirements.txt
annotated-types==0.7.0
avro==1.11.3
certifi==2024.6.2
charset-normalizer==3.3.2
grpcio==1.64.1
grpcio-tools==1.64.1
idna==3.7
protobuf==5.27.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.1
